# Optimize HuffmanArray