Various optimizations and bug fixes.
All units pass.

```
  Method |     Mean |     Error |    StdDev | Scaled | ScaledSD | Allocated |
-------- |---------:|----------:|----------:|-------:|---------:|----------:|
    Jump | 617.3 ns |  5.430 ns |  4.814 ns |   0.67 |     0.01 |       0 B |
 OrigOpt | 731.7 ns | 14.538 ns | 24.687 ns |   0.80 |     0.03 |       0 B |
    Orig | 916.1 ns | 18.082 ns | 18.569 ns |   1.00 |     0.00 |       0 B |
   Array | 426.0 ns |  2.351 ns |  2.199 ns |   0.47 |     0.01 |       0 B |
```