@jbhensley 2018-09-04
I completed integrating the two `Decode` methods, which rewrites large portions of this. I think we should concentrate on optimizing the new code.