@grant-d 2018-09-04
Sounds good, thanks for the hard work.
 
I see you skipped some of the optimizations:
- Unrolling `for (int i = 0; i < 4; i++)` and `AggressiveInline` using a `private` helper has a measurable performance improvement
- Using `short` instead of `int` drops the runtime overhead by 2x (**I am concerned** about the runtime overhead of this approach which, even using `short` and `Lazy`, is 256 x 15 x 2 bytes ~ **7.5k**
- Passing `s_decodingArray` explicitly down the call chain has a measurable performance improvement. @jcdickinson suggested this as a known JIT bug of some kind. 
- I found `Span` to be slower than raw `short[]` access
- Some other small optimizations

We can add these back incrementally and document the perf increase (if any), I will submit relevant PRs